Split contents of Issuance.kt. Was getting too big.
