+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/internal/CredentialIssuanceRequest.kt
@@ -47,7 +47,7 @@ internal sealed interface CredentialIssuanceRequest {
     ) : CredentialIssuanceRequest {
         init {
             ensure(credentialRequests.all { it.encryption == null }) {
-                BatchRequestHasEncryptionSpecInIndividualRequests
+                BatchRequestHasEncryptionSpecInIndividualRequests()
             }
         }
     }
